change data type casting in the function

+++ macros/cents_to_dollars.sql
@@ -1,3 +1,3 @@
 {%- macro cents_to_dollars(column_name, decimals=2) -%}
-    ROUND(CAST({{ column_name }} AS FLOAT) / 100.0, {{ decimals }})
+    ROUND(CAST({{ column_name }} AS NUMBER(38,2)) / 100.0, {{ decimals }})
 {% endmacro %}
